+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.2
+* Added support for GHC 8.0
+* Added `Data.Functor.Deriving` and `Data.Traversable.Deriving`, which allow deriving `Functor` and `Traversable` with TH.
+* Added `Data.Deriving`, which reexports all other modules
+
 ## 0.1
 * Initial commit